Convert filet to support parallel execution
We create a new list of files per TestReporter. This needed a map, and
thus some helper functions to help with locking around the map.

This however changes the way you can access the `Files` array. You now
need to access it through the map key (see the README changes)

Fixes #5

Signed-off-by: Jo Vandeginste <Jo.Vandeginste@kuleuven.be>